requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
packaging==25.0
pip-review==1.3.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyodbc==5.2.0
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
